net_udp.c: Fixed FIONREAD ioctl to take an int* argument instead of an unsigned long*.
git-svn-id: svn://svn.code.sf.net/p/quakespasm/code/trunk/quakespasm@502 af15c1b1-3010-417e-b628-4374ebc0bcbd

+++ Quake/net_udp.c
@@ -235,7 +235,7 @@ int UDP_Connect (sys_socket_t socketid, struct qsockaddr *addr)
 
 sys_socket_t UDP_CheckNewConnections (void)
 {
-	unsigned long	available;
+	int		available;
 	struct sockaddr_in	from;
 	socklen_t	fromlen;
 	char		buff[1];
